move STACTiler dependencies in titiler.dependencies
